Add google maps utils and tests
