+++ src/helpers.js
@@ -1,5 +1,8 @@
 import React, { useState, useEffect } from "react";
 import { ArrowLeft, ArrowRight, AlertCircle } from "lucide-react";
+import { parseGoogleMapsUrl } from "./googleMapsUtils.js";
+
+export { parseGoogleMapsUrl };
 
 /* eslint react-refresh/only-export-components: off */
 
@@ -10,109 +13,6 @@ export const getCardinalDirection = (degrees) => {
   return directions[(val % 16)];
 };
 
-// Helper function to try extracting location name from URL parameters
-const extractNameFromUrl = (url) => {
-  try {
-    // Try pattern: /place/Name/ in the URL
-    let placeMatch = url.match(/\/place\/([^\/]+)\//);
-    if (placeMatch && placeMatch[1]) {
-      const decoded = decodeURIComponent(placeMatch[1]);
-      // Clean up the name (replace + and remove coordinates)
-      return decoded
-        .replace(/\+/g, ' ')
-        .replace(/\d+\.\d+,\d+\.\d+/, '')
-        .trim();
-    }
-    
-    // Try pattern: ?q=Name in the URL
-    let qMatch = url.match(/[?&]q=([^&@]+)/);
-    if (qMatch && qMatch[1]) {
-      const decoded = decodeURIComponent(qMatch[1]);
-      // Remove coordinates if present
-      return decoded
-        .replace(/\+/g, ' ')
-        .replace(/\d+\.\d+,\d+\.\d+/, '')
-        .trim();
-    }
-    
-    return null;
-  } catch (e) {
-    console.error("Error extracting name from URL:", e);
-    return null;
-  }
-};
-
-// Helper function to try extracting location from URL parameters
-const extractLocationFromUrl = (url) => {
-  try {
-    // Try to find lat,lng pattern
-    const latLngMatch = url.match(/[?&]q=(-?\d+\.\d+),(-?\d+\.\d+)/);
-    if (latLngMatch) {
-      return {
-        lat: parseFloat(latLngMatch[1]),
-        lng: parseFloat(latLngMatch[2])
-      };
-    }
-    
-    // Try to find @lat,lng pattern
-    const atLatLngMatch = url.match(/@(-?\d+\.\d+),(-?\d+\.\d+)/);
-    if (atLatLngMatch) {
-      return {
-        lat: parseFloat(atLatLngMatch[1]),
-        lng: parseFloat(atLatLngMatch[2])
-      };
-    }
-    
-    return null;
-  } catch (e) {
-    console.error("Error extracting location from URL:", e);
-    return null;
-  }
-};
-
-// Parse Google Maps URL - standard pattern only, short URLs now handled by proxy.js
-export const parseGoogleMapsUrl = (url) => {
-  if (!url) return null;
-  
-  console.log("Parsing URL:", url);
-  
-  // Extract name if it's in the URL
-  const nameFromUrl = extractNameFromUrl(url);
-  console.log("Name extracted from URL:", nameFromUrl);
-  
-  // Handle standard Google Maps URL with @lat,lng
-  let match = url.match(/@(-?\d+\.\d+),(-?\d+\.\d+)/);
-  if (match) {
-    console.log("Coordinates extracted from URL:", match[1], match[2]);
-    return {
-      name: nameFromUrl || "New Beach",
-      latitude: parseFloat(match[1]),
-      longitude: parseFloat(match[2]),
-      googleMapsUrl: url
-    };
-  }
-  
-  // Handle ?q=lat,lng format
-  match = url.match(/\?q=(-?\d+\.\d+),(-?\d+\.\d+)/);
-  if (match) {
-    console.log("Coordinates extracted from q= parameter:", match[1], match[2]);
-    return {
-      name: nameFromUrl || "New Beach",
-      latitude: parseFloat(match[1]),
-      longitude: parseFloat(match[2]),
-      googleMapsUrl: url
-    };
-  }
-  
-  // If it's a short URL, return null to let the proxy handler take over
-  if (url.includes("goo.gl")) {
-    console.log("Short URL detected, deferring to proxy handler");
-    return null;
-  }
-  
-  console.log("No matching pattern found");
-  return null;
-};
 
 // Error Boundary Component
 export const ErrorBoundary = ({ children }) => {
